+++ game/state/battle/ai/unitaihardcore.cpp
@@ -1,0 +1,9 @@
+#include "game/state/battle/ai/unitaihardcore.h"
+
+namespace OpenApoc
+{
+
+UnitAIHardcore::UnitAIHardcore() { type = Type::Hardcore; }
+
+
+}
