+++ game/state/battle/ai/unitaihardcore.h
@@ -1,0 +1,15 @@
+#pragma once
+#include "game/state/battle/ai/unitai.h"
+
+namespace OpenApoc
+{
+	
+// AI that tries to be as hard towards the player as possible, TBD
+class UnitAIHardcore : public UnitAI
+{
+  public:
+	UnitAIHardcore();
+};
+
+
+}
