+++ game/state/gamestate.vcxproj
@@ -22,7 +22,16 @@
     <ClCompile Include="agent.cpp" />
     <ClCompile Include="base\base.cpp" />
     <ClCompile Include="base\facility.cpp" />
-    <ClCompile Include="battle\ai\ai.cpp" />
+    <ClCompile Include="battle\ai\aidecision.cpp" />
+    <ClCompile Include="battle\ai\tacticalai.cpp" />
+    <ClCompile Include="battle\ai\tacticalaivanilla.cpp" />
+    <ClCompile Include="battle\ai\unitai.cpp" />
+    <ClCompile Include="battle\ai\unitaibehavior.cpp" />
+    <ClCompile Include="battle\ai\unitaidefault.cpp" />
+    <ClCompile Include="battle\ai\unitaihardcore.cpp" />
+    <ClCompile Include="battle\ai\unitaihelper.cpp" />
+    <ClCompile Include="battle\ai\unitailowmorale.cpp" />
+    <ClCompile Include="battle\ai\unitaivanilla.cpp" />
     <ClCompile Include="battle\battle.cpp" />
     <ClCompile Include="battle\battledoor.cpp" />
     <ClCompile Include="battle\battleexplosion.cpp" />
@@ -87,7 +96,17 @@
     <ClInclude Include="agent.h" />
     <ClInclude Include="base\base.h" />
     <ClInclude Include="base\facility.h" />
-    <ClInclude Include="battle\ai\ai.h" />
+    <ClInclude Include="battle\ai\aitype.h" />
+    <ClInclude Include="battle\ai\aidecision.h" />
+    <ClInclude Include="battle\ai\tacticalai.h" />
+    <ClInclude Include="battle\ai\tacticalaivanilla.h" />
+    <ClInclude Include="battle\ai\unitai.h" />
+    <ClInclude Include="battle\ai\unitaibehavior.h" />
+    <ClInclude Include="battle\ai\unitaidefault.h" />
+    <ClInclude Include="battle\ai\unitaihardcore.h" />
+    <ClInclude Include="battle\ai\unitaihelper.h" />
+    <ClInclude Include="battle\ai\unitailowmorale.h" />
+    <ClInclude Include="battle\ai\unitaivanilla.h" />
     <ClInclude Include="battle\battle.h" />
     <ClInclude Include="battle\battlecommonsamplelist.h" />
     <ClInclude Include="battle\battledoor.h" />
