+++ game/state/gamestate.vcxproj.filters
@@ -198,10 +198,37 @@
     <ClCompile Include="battle\battlemaptileset.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="battle\ai\ai.cpp">
+    <ClCompile Include="battle\battlescanner.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="battle\battlescanner.cpp">
+    <ClCompile Include="battle\ai\aidecision.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="battle\ai\tacticalai.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="battle\ai\tacticalaivanilla.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="battle\ai\unitai.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="battle\ai\unitaihelper.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="battle\ai\unitaidefault.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="battle\ai\unitaibehavior.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="battle\ai\unitailowmorale.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="battle\ai\unitaivanilla.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="battle\ai\unitaihardcore.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
@@ -401,10 +428,40 @@
     <ClInclude Include="gamestate_serialize.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="battle\ai\ai.h">
+    <ClInclude Include="battle\battlescanner.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="battle\battlescanner.h">
+    <ClInclude Include="battle\ai\aidecision.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="battle\ai\tacticalai.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="battle\ai\tacticalaivanilla.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="battle\ai\unitai.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="battle\ai\unitaihelper.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="battle\ai\unitaidefault.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="battle\ai\unitaibehavior.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="battle\ai\unitailowmorale.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="battle\ai\unitaivanilla.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="battle\ai\unitaihardcore.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="battle\ai\aitype.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
