+++ game/state/gamestate_serialize.h
@@ -9,6 +9,13 @@
 #include "game/state/agent.h"
 #include "game/state/base/base.h"
 #include "game/state/base/facility.h"
+#include "game/state/battle/ai/aitype.h"
+#include "game/state/battle/ai/tacticalaivanilla.h"
+#include "game/state/battle/ai/unitaivanilla.h"
+#include "game/state/battle/ai/unitaidefault.h"
+#include "game/state/battle/ai/unitailowmorale.h"
+#include "game/state/battle/ai/unitaibehavior.h"
+#include "game/state/battle/ai/unitaihardcore.h"
 #include "game/state/battle/battle.h"
 #include "game/state/battle/battlecommonimagelist.h"
 #include "game/state/battle/battlecommonsamplelist.h"
